Tested Build with Supposedly Faulty Script
TapToMakeAppear.cs was thought to be a cause of the build not working.  I added Placement Manager back in and the build works.  It does not prevent MRTK from working, but it is recommended to find a way to toggle off TTP while in Add mode and toggle TTP on while in Move mode.

Rotate mode is useless and should be removed or replaced with a removal option.